xhk: Check mirrored status before changing state
Sometimes, a user can while typing normally overlap a 'non-mirroring' key such as 'shift' or 'control'
while the space key is held. This can cause an unnecessary transition to the space-modified state
and will cause us to lose the space bar press even though no mirroring was desired.

This can often happen when a space is followed by a shift key in preparation of a capital letter.